MOB-217 Test multipaxos integration
Added option for testing with multipaxos package, and did other fixes for demultiplexing, proposing requests, and handling decisions. The option is enabled by '-multipaxos' at name servers.  
multipaxos does not work with GNS correctly.

git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@854 0b15c3b6-4604-4ca4-ad59-2e1af984e93c

+++ src/edu/umass/cs/gns/paxos/AbstractPaxosManager.java
@@ -36,7 +36,7 @@ public abstract class AbstractPaxosManager {
    */
   public abstract String propose(String paxosIDNoVersion, String value);
 
-  public abstract String proposeStop(String paxosIDNoVersion, String value, int version);
+  public abstract String proposeStop(String paxosIDNoVersion, String value, short version);
 
 
   /**
